moved the getAllRecipes to RecipeCatalog.tsx with it useEffect
