+++ frontend/src/App.tsx
@@ -1,30 +1,8 @@
-
 import './App.css'
-import {useEffect, useState} from "react";
-import type {Recipe} from "./Recipe.ts";
 import NavigationLinks from "./navigation/NavigationLinks.tsx";
-import axios from "axios";
 import RoutingPaths from "./navigation/RoutingPaths.tsx";
 
 export default function App() {
-
-    const [recipes, setRecipes] = useState<Recipe[]>()
-
-    function getAllRecipes() {
-        axios.get("api/recipes")
-            .then(Response => setRecipes(Response.data))
-            .catch((error) => console.log("Function: getAllRecipes. ERROR: " + error))
-    }
-
-    useEffect(() => {
-        getAllRecipes()
-    }, []);
-
-    if (!recipes)
-    {
-        return "loading recipes ..."
-    }
-
   return (
       <>
           <div className="grid">
@@ -35,7 +13,7 @@ export default function App() {
                   <NavigationLinks/>
               </div>
               <div className={"window-pane right"}>
-                  <RoutingPaths recipes={recipes}/>
+                  <RoutingPaths/>
               </div>
           </div>
       </>
