+++ frontend/src/components/RecipeCatalog.tsx
@@ -1,15 +1,32 @@
 
 import type {Recipe} from "../Recipe.ts";
 import RecipeCard from "./RecipeCard.tsx";
+import {useEffect, useState} from "react";
+import axios from "axios";
 
-type Props = {
-    recipes: Recipe[],
-}
-export default function RecipeCatalog(props:Props) {
+export default function RecipeCatalog() {
+
+    const [recipes, setRecipes] = useState<Recipe[]>()
+
+    function getAllRecipes() {
+        axios.get("api/recipes")
+            .then(Response => setRecipes(Response.data))
+            .catch((error) => console.log("Function: getAllRecipes. ERROR: " + error))
+    }
+
+    useEffect(() => {
+        getAllRecipes()
+    }, []);
+
+    if (!recipes)
+    {
+        return "loading recipes ..."
+    }
+
     return (
         <>
             {
-                props.recipes.map(
+                recipes.map(
                     (recipe:Recipe)=> <RecipeCard key={recipe.id} recipe={recipe}/>
                 )
             }
